Refactor rating issues
