+++ client/app/intake/pages/supplementalClaim/finish.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { Redirect } from 'react-router-dom';
 import Button from '../../../components/Button';
 import CancelButton from '../../components/CancelButton';
-import RatedIssueCounter from '../../components/RatedIssueCounter';
+import { RatedIssuesUnconnected, RatedIssueCounter } from '../../components/RatedIssues';
 import NonRatedIssues from './nonRatedIssues';
-import RatedIssues from './ratedIssues';
 import { connect } from 'react-redux';
 import { completeIntake } from '../../actions/supplementalClaim';
+import { setIssueSelected } from '../../actions/common';
 import { REQUEST_STATE, PAGE_PATHS, INTAKE_STATES } from '../../constants';
 import { bindActionCreators } from 'redux';
 import { getIntakeStatus } from '../../selectors';
@@ -53,6 +53,16 @@ class Finish extends React.PureComponent {
     </div>;
   }
 }
+
+const RatedIssues = connect(
+  ({ supplementalClaim, intake }) => ({
+    intakeId: intake.id,
+    reviewState: supplementalClaim
+  }),
+  (dispatch) => bindActionCreators({
+    setIssueSelected
+  }, dispatch)
+)(RatedIssuesUnconnected);
 
 class FinishNextButton extends React.PureComponent {
   handleClick = () => {
